deepen: dissolve sec-section-extractor into the SEC Section module
Replaces two shallow modules around one CONTEXT.md concept ("SEC
Section") with one deep module behind a five-function interface.

Before:
  lib/parsers/sec-section-extractor.ts  (exports extractSections,
                                         hasUsableFinancialSection;
                                         imports ExtractedSection from
                                         context-manager)
  lib/ai/prompts/context-manager.ts     (exports ExtractedSection,
                                         buildSectionedPrompt,
                                         SECTION_TOKEN_BUDGETS,
                                         SECTION_PRIORITY)

The split was load-bearing for nothing: sec-section-extractor's only
production caller (`lib/ai/summarize.ts`) used both modules in
lockstep, and sec-section-extractor depended on `ExtractedSection`
defined in context-manager — a textbook sign that the two halves
belonged in one module. The seam between them had one adapter and one
caller, the very pattern LANGUAGE.md flags ("one adapter means a
hypothetical seam").

After: `lib/ai/prompts/context-manager.ts` is the canonical SEC Section
module. The interface is:

  extractSections(cleanedText, formType) → ExtractedSection[]
  hasUsableFinancialSection(sections)     → boolean
  buildSectionedPrompt(sections, budget) → string
  SECTION_TOKEN_BUDGETS                   (per-section budgets)
  SECTION_PRIORITY                        (priority-ordered ordering)

The `HEADING_TO_SECTION` table and the `matchHeadingToSection`
internal helper are now private to the module — callers no longer
branch on heading shape or form type.

Dependency category: in-process (pure computation, no I/O), so the
deepening is a straight inline — no port, no adapter.

Tests: 19 tests at the deepened interface (renamed to
`__tests__/ai/prompts/sec-section-module.test.ts` to reflect that the
test surface IS the module). All 19 pass. No new tests added; the
existing extractor tests already asserted on observable outcomes
through the interface, so they survived the move unchanged.

CONTEXT.md "SEC Section" entry updated to reflect the consolidated
interface and the deletion of the parallel extractor module.
