+++ CONTEXT.md
@@ -31,7 +31,7 @@ An enrichment of a [Summary] that captures public discussion on X (Twitter) abou
 A property of a [Filing] meaning the SEC document body, after fetching and cleaning, contains enough structured financial-statement signal for the AI to produce a usable [Summary]. After the NVDA 10-Q fix shipped in this branch, "extractable" has a precise definition: the filing passes both `hasFinancialStatementSignal()` pre-LLM (3 of 5 signals present in the cleaned text: period header, statement title, line items, dollar density, minimum length — see `lib/ai/parsers/financial-content-gate.ts`) AND `hasUsableFinancialHighlights()` post-LLM (at least one `financialHighlights` row passes `isUsableMetricRow`). When the user-facing email says "no extractable financial results," it means one of these gates failed. Filings where `requiresFinancialContent(formType)` is true (10-K, 10-Q, 20-F, 6-K, and amendments) MUST be extractable; failure is treated as a parser bug and the email is not sent. Filings outside that set (Form 4, NT-10-K, etc.) may legitimately be non-extractable and fall through to the minimal-content metadata path.
 
 ### SEC Section
-A semantic region of a [Filing] — "Risk Factors", "Management Discussion", "Financial Statements", "Business Overview", etc. Canonical type: `SECFilingSection` in `lib/ai/prompts/prompt-types.ts`. Per-section token budgets and (forthcoming) priority order live alongside that type in `lib/ai/prompts/context-manager.ts`. Anything that segments filing text into these buckets, or assigns budgets to them, must reuse `SECFilingSection` — do not define parallel section vocabularies. Not to be confused with `FilingSection` in `lib/parsers/html-parser.ts`, which is a different concept: structural HTML element types (TITLE, HEADER, PARAGRAPH, TABLE, LIST) used only by the standalone html-parser module.
+A semantic region of a [Filing] — "Risk Factors", "Management Discussion", "Financial Statements", "Business Overview", etc. Canonical type: `SECFilingSection` in `lib/ai/prompts/prompt-types.ts`. Implemented as one deep module (`lib/ai/prompts/context-manager.ts`) behind a five-function interface — `extractSections(cleanedText, formType)` segments cleaned markdown into the canonical buckets, `hasUsableFinancialSection(sections)` is the post-extract gate the Summarize module uses to decide whether to throw AI_INSUFFICIENT_FINANCIAL_SECTION, `buildSectionedPrompt(sections, budget)` assembles a priority-ordered prompt body honoring per-section token budgets, and the `SECTION_TOKEN_BUDGETS` + `SECTION_PRIORITY` tables live alongside. Form-type-aware heading routing (10-Q Item 1 = Financial Statements vs 10-K Item 1 = Business Overview, and 10-Q Part II Item 1 = Legal Proceedings vs 10-Q Part I Item 1 = Financial Statements) is private to the module's `HEADING_TO_SECTION` table — callers never branch on form-type or heading shape. Replaces `lib/parsers/sec-section-extractor.ts`, which split the heading-routing table from the priority-ordered prompt assembly even though it depended on `ExtractedSection` from context-manager and had exactly one production caller (`lib/ai/summarize.ts`); the split was a shallow seam (one adapter, one caller, no leverage). Anything that segments filing text into these buckets, or assigns budgets to them, must reuse `SECFilingSection` — do not define parallel section vocabularies. Not to be confused with `FilingSection` in `lib/parsers/html-parser.ts`, which is a different concept: structural HTML element types (TITLE, HEADER, PARAGRAPH, TABLE, LIST) used only by the standalone html-parser module.
 
 ### Lifetime Seat
 A user who claimed one of the first 25 lifetime seats via a one-time $499 payment. Entitled to all MAX-tier features (see `lib/stripe/plans.ts`) forever, with no recurring subscription. Modeled as a `User.foundingMember: Boolean` flag rather than a new `SubscriptionTier` value, since the entitlement set is identical to MAX. Purchased via a one-time Stripe `Price` (env: `STRIPE_FOUNDING_LIFETIME_PRICE_ID`) under the existing MAX `Product`, not a recurring subscription. Not to be confused with [Subscription] (watchlist interest) or `UserSubscription` (Stripe-side recurring plan record, which Lifetime Seat holders do not have).
